% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ol1.R
\name{ol1}
\alias{ol1}
\title{A function to estimate coefficient vector beta}
\usage{
ol1(XXmat, cvec, inival, lambda, varset, maxiter, tol)
}
\arguments{
\item{XXmat}{X is columnwise rescaled to make each column have L2 norm root-n and XXmat=n^(-1)t(X)X}

\item{cvec}{n^(-1)t(X)y}

\item{inival}{initial value of beta}

\item{lambda}{ridge parameter}

\item{varset}{variance set}

\item{maxiter}{maximum iteration time}

\item{tol}{limitation to end the iteration}
}
\value{
etimator of beta
}
\description{
min_beta  (2 Lam)^(-1) (z - beta)^2 + p_lambda(|beta|) for L1 penalty
}
