\name{SC19084-package}
\alias{SC19084-package}
\alias{SC19084}
\docType{package}
\title{
  High dimensional thresholded regression.
}
\description{
  We intended to implement the high dimensional threshold regression applying ICA algorithm. In this package£¬the ol1 function was built to estimate coefficient vector beta for L1 penalty and the uhard function was constructed as one step to realise ICA algorithm which is applyed to implement the  high dimensional threshold regression. Therefore, we can compare the performance of the two methods under circumstance of high dimensional regression.
}
\author{
Sherry J

Maintainer: Sherry J <xyz19z@mail.ustc.edu.cn>
}
\keyword{ package }
\examples{
  \dontrun{
     ## Optional simple examples of the most important functions
     ## These can be in \dontrun{} and \donttest{} blocks.   
  }
}
