% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uhard.R
\name{uhard}
\alias{uhard}
\title{uhard}
\usage{
uhard(z, Lam, lambda0, lambda)
}
\arguments{
\item{z}{wavelet transform coefficient}

\item{Lam}{control coefficieent}

\item{Lamda0}{for initial value of z}

\item{Lamda}{threshold value}
}
\value{
beta for hard-thresholding penalty
}
\description{
min_beta  (2 Lam)^(-1) (z - beta)^2 + lambda0 |beta| + p_lambda(|beta|) for hard-thresholding penalty
}
